Fix printing bugs

+++ server/data/MTGCardData.py
@@ -108,7 +108,7 @@ class MTGCardData(CardData):
 
                 art = art.resize((self.IMAGE_WIDTH, new_height))
                 art = ImageOps.grayscale(art)
-                art = art.point(lambda x: 255 - (255 - x)/2)
+                art = art.point(lambda x: 255 - int((255 - x)/2))
 
                 if not os.path.isdir(self.art_folder):
                     os.makedirs(self.art_folder)
